Extract manual apply live server module
